+++ CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * Allow any `ServiceResetterInterface` implementation in `ResetServicesListener`
  * Add `Symfony\Component\Messenger\Middleware\AddDefaultStampsMiddleware` and `Symfony\Component\Messenger\Message\DefaultStampsProviderInterface`
  * Add the possibility to configure exchange to exchange bindings in AMQP transport
+ * Add `MessageSentToTransportsEvent` that is dispatched only after the message was sent to at least one transport
 
 7.3
 ---